# Faster validations + record reference stability at the same time