--- a/packages/editor/src/lib/utils/SharedStylesMap.ts
+++ b/packages/editor/src/lib/utils/SharedStylesMap.ts
@@ -35,7 +35,7 @@ function sharedStyleEquals<T>(a: SharedStyle<T>, b: SharedStyle<T> | undefined) 
  */
 export class ReadonlySharedStyleMap {
 	/** @internal */
-	protected map: Map<StyleProp<unknown>, SharedStyle<unknown>>
+	protected map: Map<StyleProp<any>, SharedStyle<unknown>>
 
 	constructor(entries?: Iterable<[StyleProp<unknown>, SharedStyle<unknown>]>) {
 		this.map = new Map(entries)