--- a/packages/store/src/lib/RecordType.ts
+++ b/packages/store/src/lib/RecordType.ts
@@ -29,7 +29,7 @@ export class RecordType<
 > {
 	readonly createDefaultProperties: () => Exclude<OmitMeta<R>, RequiredProperties>
 	readonly migrations: Migrations
-	readonly validator: StoreValidator<R> | { validate: (r: unknown) => R }
+	readonly validator: StoreValidator<R>
 
 	readonly scope: RecordScope
 
@@ -44,7 +44,7 @@ export class RecordType<
 		config: {
 			readonly createDefaultProperties: () => Exclude<OmitMeta<R>, RequiredProperties>
 			readonly migrations: Migrations
-			readonly validator?: StoreValidator<R> | { validate: (r: unknown) => R }
+			readonly validator?: StoreValidator<R>
 			readonly scope?: RecordScope
 		}
 	) {
@@ -198,7 +198,10 @@ export class RecordType<
 	 * Check that the passed in record passes the validations for this type. Returns its input
 	 * correctly typed if it does, but throws an error otherwise.
 	 */
-	validate(record: unknown): R {
+	validate(record: unknown, recordBefore?: R): R {
+		if (recordBefore && this.validator.validateUsingKnownGoodVersion) {
+			return this.validator.validateUsingKnownGoodVersion(recordBefore, record)
+		}
 		return this.validator.validate(record)
 	}
 }