@steveruizok 2024-02-16
Nice! I like the idea of validating from a previous "known good version".